Controller 수정 -refactor: Controller의 비즈니스 로직 Model영역으로 옮김. -refactor: Model영역에 GameHelper class를 추가해 비즈니스 로직을 구현 -refactor: Outputview 수정
